django_app/requirements.txt
django
psycopg2-binary
djangorestframework
django-cors-headers
pandas
numpy  
scikit-learn  
scipy